requirements.txt
fastmcp-http
wikipedia
openai
python-dotenv
requests
